Final refactor of cAdvisor config for unsupported platforms. Pared down stub integrations.
